Reorganizing quicksort
