--- quicksort_algo/.vscode/launch.json
@@ -1,13 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "C/C++ Runner: Debug Session",
-      "type": "lldb",
-      "request": "launch",
-      "args": [],
-      "cwd": "/Users/treymiller/Education/cpp_knowledge/useful_algos/quicksort_algo",
-      "program": "/Users/treymiller/Education/cpp_knowledge/useful_algos/quicksort_algo/build/Debug/outDebug"
-    }
-  ]
-}
