+++ quicksort_algo/include/quicksort_algo.hpp
@@ -1,0 +1,11 @@
+#include <iostream>
+#include <vector>
+#include <string>
+#include <map>
+#include <cmath>
+
+// TODO: turn these two functions into templates so that we can use more variable types
+// in them. 
+
+void partitionParent(std::vector<int> &arr, int low, int high);
+int partition(std::vector<int>& arr, int& low, int& high);
